# swagger.json
{
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This package provides a REST API to the Ignition Fuel server.",
    "title": "Ignition Fuel Server REST API",
    "contact": {
      "email": "info@openrobotics.org"
    },
    "license": {
      "name": "Apache 2.0"
    },
    "version": "0.1.0"
  },
  "host": "fuel.ignitionrobotics.org",
  "basePath": "/1.0",
  "paths": {
...
    "/models": {
      "get": {
        "description": "Get a list of models. Models will be returned paginated,\nwith pages of 20 models by default. The user can request a\ndifferent page with query parameter 'page', and the page size\ncan be defined with query parameter 'per_page'.\nThe route supports the 'order' parameter, with values 'asc' and\n'desc' (default: desc).\nIt also supports the 'q' parameter to perform a fulltext search on models\nname, description and tags.",
        "produces": [
          "application/json",
          "application/x-protobuf"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "models"
        ],
        "summary": "Get list of models.",
        "operationId": "listModels",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "SearchQuery",
            "description": "Search query",
            "name": "q",
            "in": "query"
          },
          {
            "enum": [
              "asc",
              " desc"
            ],
            "type": "string",
            "default": "desc",
            "x-go-name": "Order",
            "name": "order",
            "in": "query"
          },
          {
            "minimum": 1,
            "type": "integer",
            "format": "int64",
            "default": 1,
            "x-go-name": "Page",
            "description": "The page to return",
            "name": "page",
            "in": "query"
          },
          {
            "maximum": 100,
            "minimum": 1,
            "type": "integer",
            "format": "int64",
            "default": 20,
            "x-go-name": "PageSize",
            "description": "Size of the pages",
            "name": "per_page",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/jsonModels"
          },
          "default": {
            "$ref": "#/responses/fuelError"
          }
        }
      },
      "post": {
        "description": "Creates a new model. The request body should contain the\nfollowing fields: 'modelName', 'urlName', 'description',\n'license' (number), 'permission' (number). All values as strings.\n'tags': a string containing a comma separated list of tags.\nThe model owner will be retrieved from the passed JWT.\n'file': multiple files in the multipart form.",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "models"
        ],
        "summary": "Create model",
        "operationId": "createModel",
        "parameters": [
          {
            "x-go-name": "Model",
            "description": "Model data",
            "name": "model",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "dbModel",
            "schema": {
              "$ref": "#/definitions/dbModel"
            }
          },
          "default": {
            "$ref": "#/responses/fuelError"
          }
        }
      }
    },
...
    "/worlds": {
      "get": {
        "description": "Get a list of worlds. Worlds will be returned paginated,\nwith pages of 20 worlds by default. The user can request a\ndifferent page with query parameter 'page', and the page size\ncan be defined with query parameter 'per_page'.\nThe route supports the 'order' parameter, with values 'asc' and\n'desc' (default: desc).\nIt also supports the 'q' parameter to perform a fulltext search on worlds\nname, description and tags.",
        "produces": [
          "application/json",
          "application/x-protobuf"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "worlds"
        ],
        "summary": "Get list of worlds.",
        "operationId": "listWorlds",
        "responses": {
          "200": {
            "$ref": "#/responses/jsonWorlds"
          },
          "default": {
            "$ref": "#/responses/fuelError"
          }
        }
      },
      "post": {
        "description": "Creates a new world. The request body should contain the\nfollowing fields: 'name', 'description',\n'license' (number), 'permission' (number). All values as strings.\n'tags': a string containing a comma separated list of tags.\nThe worlds owner will be retrieved from the passed JWT.\n'file': multiple files in the multipart form.",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "worlds"
        ],
        "summary": "Create world",
        "operationId": "createWorld",
        "responses": {
          "200": {
            "description": "dbWorld",
            "schema": {
              "$ref": "#/definitions/dbWorld"
            }
          },
          "default": {
            "$ref": "#/responses/fuelError"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Categories": {
      "description": "Categories is an array of Category",
      "type": "array",
      "items": {
        "$ref": "#/definitions/Category"
      },
      "x-go-package": "github.com/gazebo-web/fuel-server/bundles/category"
    },
...
    "CreateModel": {
      "description": "CreateModel encapsulates data required to create a model",
      "type": "object",
      "required": [
        "name",
        "license",
        "file"
      ],
      "properties": {
        "categories": {
          "description": "Categories",
          "type": "string",
          "maximum": 2,
          "x-go-name": "Categories"
        },
        "description": {
          "description": "Optional description",
          "type": "string",
          "x-go-name": "Description"
        },
        "file": {
          "description": "One or more files",
          "type": "string",
          "x-go-name": "File"
        },
        "license": {
          "description": "License ID",
          "type": "integer",
          "format": "int64",
          "minimum": 1,
          "x-go-name": "License"
        },
        "metadata": {
          "$ref": "#/definitions/ModelMetadata"
        },
        "name": {
          "description": "The name of the Model",
          "type": "string",
          "x-go-name": "Name"
        },
        "owner": {
          "description": "Optional Owner of the model. Must be a user or an org.\nIf not set, the current user will be used as owner",
          "type": "string",
          "x-go-name": "Owner"
        },
        "permission": {
          "description": "The associated permissions. 0 for public, 1 for private models.",
          "type": "integer",
          "format": "int64",
          "enum": [
            0,
            " 1"
          ],
          "x-go-name": "Permission"
        },
        "private": {
          "description": "Optional privacy/visibility setting.",
          "type": "boolean",
          "x-go-name": "Private"
        },
        "tags": {
          "description": "A comma separated list of tags",
          "type": "string",
          "x-go-name": "Tags"
        },
        "urlName": {
          "description": "Url name",
          "type": "string",
          "x-go-name": "URLName"
        }
      },
      "x-go-package": "github.com/gazebo-web/fuel-server/bundles/models"
    },
...
    "ErrMsg": {
      "description": "ErrMsg is serialized as JSON, and returned if the request does not succeed\nTODO: consider making ErrMsg an 'error'",
      "type": "object",
      "properties": {
        "errcode": {
          "description": "Internal error code.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrCode"
        },
        "errid": {
          "description": "Generated ID for easy tracking in server logs",
          "type": "string",
          "x-go-name": "ErrID"
        },
        "extra": {
          "description": "Extra information/arguments associated to Error message.",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Extra"
        },
        "msg": {
          "description": "Error message.",
          "type": "string",
          "x-go-name": "Msg"
        },
        "remote-address": {
          "description": "Associated request remote address, if applicable",
          "type": "string",
          "x-go-name": "RemoteAddress"
        },
        "route": {
          "description": "Associated request Route, if applicable",
          "type": "string",
          "x-go-name": "Route"
        },
        "user-agent": {
          "description": "Associated request User-Agent, if applicable",
          "type": "string",
          "x-go-name": "UserAgent"
        }
      },
      "x-go-package": "gitlab.com/ignitionrobotics/web/ign-go"
    },
...
    "License": {
      "description": "License is a license name and ID",
      "type": "object",
      "properties": {
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "DeletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "ID": {
          "type": "integer",
          "format": "uint64"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "image_url": {
          "type": "string",
          "x-go-name": "ImageURL"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "url": {
          "type": "string",
          "x-go-name": "ContentURL"
        }
      },
      "x-go-package": "github.com/gazebo-web/fuel-server/bundles/license"
    },
...
    "Model": {
      "description": "Model base model definition, including fields `ID`, `CreatedAt`, `UpdatedAt`, `DeletedAt`, which could be embedded in your models\ntype User struct {\ngorm.Model\n}",
      "type": "object",
      "properties": {
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "DeletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "ID": {
          "type": "integer",
          "format": "uint64"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "github.com/jinzhu/gorm"
    },
...
    "ModelMetadata": {
      "description": "Metadata is an array of Metadatum",
      "type": "array",
      "items": {
        "$ref": "#/definitions/dbModel"
      },
      "x-go-package": "github.com/gazebo-web/fuel-server/bundles/models"
    },
...
    "Tags": {
      "description": "Tags is an array of Tag",
      "type": "array",
      "items": {
        "$ref": "#/definitions/Tag"
      },
      "x-go-package": "github.com/gazebo-web/fuel-server/bundles/models"
    },
...
    "dbModel": {
      "description": "A model contains information about a single simulation object, such\nas a robot, table, or structure.",
      "type": "object",
      "title": "Model represents information about a simulation model",
      "properties": {
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "DeletedAt": {
          "description": "Added 2 milliseconds to DeletedAt field, and added it to the unique index to help disambiguate\nwhen soft deleted rows are involved.",
          "type": "string",
          "format": "date-time"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "categories": {
          "$ref": "#/definitions/Categories"
        },
        "creator": {
          "description": "The username of the User that created this model (usually got from the JWT)",
          "type": "string",
          "x-go-name": "Creator"
        },
        "description": {
          "description": "A description of the model (max 65,535 chars)\nInteresting post about TEXT vs VARCHAR(30000) performance:\nhttps://nicj.net/mysql-text-vs-varchar-performance/",
          "type": "string",
          "x-go-name": "Description"
        },
        "downloads": {
          "description": "Number of downloads",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Downloads"
        },
        "filesize": {
          "description": "Bytes of the model, when downloaded as a zip",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Filesize"
        },
        "lic_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "LicenseID"
        },
        "license": {
          "$ref": "#/definitions/License"
        },
        "likes": {
          "description": "Number of likes",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Likes"
        },
        "metadata": {
          "$ref": "#/definitions/ModelMetadata"
        },
        "modify_date": {
          "description": "Date and time the model was modified",
          "type": "string",
          "format": "date-time",
          "x-go-name": "ModifyDate"
        },
        "name": {
          "description": "The name of the model",
          "type": "string",
          "x-go-name": "Name"
        },
        "owner": {
          "description": "The owner of this model (must exist in UniqueOwners). Can be user or org.",
          "type": "string",
          "x-go-name": "Owner"
        },
        "permission": {
          "description": "Permission - 0: public, 1: owner, (future: team, others)",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Permission"
        },
        "private": {
          "description": "Private - True to make this a private resource",
          "type": "boolean",
          "x-go-name": "Private"
        },
        "tags": {
          "$ref": "#/definitions/Tags"
        },
        "upload_date": {
          "description": "Date and time the model was first uploaded",
          "type": "string",
          "format": "date-time",
          "x-go-name": "UploadDate"
        },
        "url_name": {
          "description": "Optional user friendly name to use in URLs",
          "type": "string",
          "x-go-name": "URLName"
        }
      },
      "x-go-name": "Model",
      "x-go-package": "github.com/gazebo-web/fuel-server/bundles/models"
    },
    "dbWorld": {
      "type": "object",
      "title": "World represents information about a simulation world.",
      "properties": {
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "DeletedAt": {
          "description": "Added 2 milliseconds to DeletedAt field, and added it to the unique index\nto help disambiguate when soft deleted rows are involved.",
          "type": "string",
          "format": "date-time"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "creator": {
          "description": "The username of the User that created this world (usually got from the JWT)",
          "type": "string",
          "x-go-name": "Creator"
        },
        "description": {
          "description": "A description of the world (max 65,535 chars)\nInteresting post about TEXT vs VARCHAR(30000) performance:\nhttps://nicj.net/mysql-text-vs-varchar-performance/",
          "type": "string",
          "x-go-name": "Description"
        },
        "downloads": {
          "description": "Number of downloads",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Downloads"
        },
        "filesize": {
          "description": "Bytes of the world, when downloaded as a zip",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Filesize"
        },
        "lic_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "LicenseID"
        },
        "license": {
          "$ref": "#/definitions/License"
        },
        "likes": {
          "description": "Number of likes",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Likes"
        },
        "modify_date": {
          "description": "Modification Date and time",
          "type": "string",
          "format": "date-time",
          "x-go-name": "ModifyDate"
        },
        "name": {
          "description": "The name of the world",
          "type": "string",
          "x-go-name": "Name"
        },
        "owner": {
          "description": "The user who created this world",
          "type": "string",
          "x-go-name": "Owner"
        },
        "permission": {
          "description": "Permission - 0: public, 1: owner, (future: team, others)",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Permission"
        },
        "private": {
          "description": "Private - True to make this a private resource",
          "type": "boolean",
          "x-go-name": "Private"
        },
        "tags": {
          "$ref": "#/definitions/Tags"
        },
        "upload_date": {
          "description": "Date and time the world was first uploaded",
          "type": "string",
          "format": "date-time",
          "x-go-name": "UploadDate"
        }
      },
      "x-go-name": "World",
      "x-go-package": "github.com/gazebo-web/fuel-server/bundles/worlds"
    }
  },
  "responses": {
...
    "fuelError": {
      "description": "Ign Fuel error serialized as JSON",
      "schema": {
        "$ref": "#/definitions/ErrMsg"
      }
    },
    "jsonModels": {
      "description": "Array of Models",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Model"
        }
      }
    }
  }
}